[virtualization][tests] Reduce the virtio block tests
Instead of reading and writing from a scattering of offsets throughout
the block devices, we can just use a starting offset, an ending offset,
and an arbitrary offset in the middle.

Also slightly modify the QCOW write tests to run over a mapped cluster,
this is just slightly more convenient.

This change saves about 12 seconds over all the block tests, so it's not
a significant reduction but there's no reason not to do it.

Change-Id: I6328339de27d88caa46c71519594137b608debf2