[intel-hda] Ensure audio can be transferred to outputs.
Attempt to play silence to each output device.

We can't actually ensure that audio was produced on the speaker, but we
can exercise the ring buffer, DMA, etc etc. Failures in these areas will
cause either an error to be produced by the Play() method, or the stream
to just halt without making progress.

Change-Id: I52a9675690791714e91375781182327270a6fd82