[audio] Fix AudioOutput streams incorrectly marked as input.
AudioOutput streams created from a device under "/dev" are incorrectly
marked as an input. This causes the buffer to be mapped into memory
as read only, in turn causing faults when the AudioOutput object
attempts to memset the buffer to zero.

This CL updates it to be marked as an output.

Tests: Intel HDA tests now use device path, instead of device number.

Change-Id: I8f5bbbaee87499d83dcf777a275e49f6dc93e24f