[Setup] Reboot after setup network interfaces change.
This changelist initiates a reboot after the network interfaces used for
setup changes so that the change can be properly recognized during
setup.

Fixes: fxb/37157
Test: fx run-test setui_service_tests -- --test
Testing description: The setup tests have been extended to use a
fake DeviceAdmin service that records requests. After the interface
change, the test verifies that the DeviceAdmin service was indeed
invoked for a reboot action.

Change-Id: Icde81ac36388f33e25fc6bfa7ce6a94babbc9636