[debugger] Implement shift.
Implements shift left and shift right mathematical operators.

Shift right is disambiguated from the same character sequence in a
template by the parser, the tokenizer always tokenizes ">>" as two
separate "greater" tokens. Shift left is handled normally since it's not
ambiguous.

This requires special-casing in the parser to put the two tokens batk
together in the cases where an operator is expected.

BUG=6398

Change-Id: I0895040c09564203d0dea8ca28d72dcdf20b9c36