[docs] update zxcrypt.md to remove outdated warnings
Several of the caveats contained therein have been resolved since that
doc was originally written. ZX-1130 has been closed, so no need to
screech about it being insecure -- security (or lack thereof) is not
a property of zxcrypt itself, but of the integration of key management
at the product level.

Bug: 32519
Change-Id: Iaf781e0e3937c8a6dc9fd905a56b014e7f962b5b