[debugger] Show source code in disassembly.
Outputs source information in disassembly dumps when available. This
makes it much easier to debug machine code.

Adds an abstraction for getting source files so the new code could be
tested with mock file contents. This has the effect of cleaning up some
of the calls because it can replace the ugly vector of directories with
an encapsulated object.

BUG=5354

Change-Id: I0724c0d26162ec81c674bacd98bae4d750ef860b