[fonts] Font manifest generator
Implementation of font_manifest_generator, a host tool that combines
font metadata from a recipe script, a developer-curated catalog of
fonts, GN variables, and the font files themselves to generate a font
manifest to be used by the font server.

Includes unit test coverage for smaller components and a larger
integration test.

Bug: 8892
Change-Id: I14a9d00cae5bec2c021e575839f6a58034830eca