[run_test_component] Allow fuchsia.boot.FactoryItems system service
The fuchsia.boot.FactoryItems protocol is served by bootsvc to provide
access to factory provisioned device data. In some cases, this data can
be faked for test usage. However, there are some cases where faked data
is insufficient and the test requires the real data. An example of this
would be a file that was encrypted with a device unit specific secret.

This CL also updates the corresponding documentation and unit test to
match the current list of allowed system services. There were several
protocols that were not listed in either.

Change-Id: I2517179af658cd6f06694d4261b70f48cf9675e8