[debugger] Convert to new index.
This keeps the old file names and forwards the includes with a shim so
it's clear what changed. The *2 files will be renamed in a followup.

There should be no observable change in behavior.

The main changes are in find_name* which has to deal with the new index
format, the fact that namespaces no longer save DIEs, and the fact that
the index walker can produce multiple nodes, each containing multiple
maps.

A test for namespace finding is added which was under-tested before.

There are minor updates in verbs_symbol.cc and index_test_support.cc.
The rest of the changes are trivial renames.

Change-Id: I25a8b0ee10b42c642c07927bc0f9764c94cdf7c6