[zircon][sysinfo] Initialize a char array in sysinfo-test
This addresses the failing test /boot/test/sys/sysinfo-test when
automatic variable initialization is enabled by zero-initializing a char
array used in the test. The test fails from a check to strlen() since
the char array it checks was not null terminated with auto-var-init
enabled and was caught by ASan.

This should be the last failing test that needs to be addressed before
enabling auto-vat-init.

Also update the test to use std::array instead of a raw array, and just
check that actual_size > 0.

BUG: 3261
Change-Id: I3f3913a8e18c94a736ad8d928b93b79b8ef52a6d