[Zircon] Use -fvisibility=hidden by default in zircon
This CL adds `-fvisibility=hidden` to all zircon targets by default.

For each target that contained a symbol which appeared in an `undefined
symbol` linker error, we simply remove this config and allow the symbols
to have default visibility. We will go back later to remove instances
where we do this by explicitely marking these undefined symbols with
`__EXPORT` (as in
https://fuchsia-review.googlesource.com/c/fuchsia/+/311763).

This CL also removes individual instances where `visibility_hidden` is
used to prevent duplicate items in the same list.

Bug: 31763
Change-Id: I9dcb98cebcc90ca077548b0ee93c9e160fc504e1