[kernel] Explicit fault handling in aspace enumeration
Uses the _capture_faults user copy variants to manually handle page
faults outside of the aspace enumeration. Handling outside of the
enumeration means the aspace lock is no longer held and we do not
have to worry about whether we are copying into the same aspace or not.

With this change the not own process restriction on the object_get_info
syscall can be relaxed.

Bug: 34427

Change-Id: If12c25f94c0d7e008f8fbdeefc007bc576d8082b