[ta] deploy TAs through ZBI 1
This is the 1st CL to deploy TAs through ZBI.
In the next CLs, the TA declaration will be moved to device-specific .gni files,
and declarations in zircon/system/dev/tee/optee/BUILD.gn will be removed.

Also selectively deploy TAs to Fuchsia and Zedboot to save space.
Will check in after fxr/319660

Bug: 36561
Test: locally tested TAs deployed correctly
Change-Id: Ifc7e06c673fe2ec180f8c7e3a3e63197169e9170