[host-tools] move directory to avoid label clobber
As we now have //tools, the tools output now creates collisions between the
group labels produced by GN and output targets that match a source directory
name produced by install_host_tools. We move the host tools to host-tools,
which is only sidestepping the issue, but should give us time to decide on a
better plan.

Change-Id: I1d92b59ed275002701ba9ac4633baba3847d3002