[docs][fuzzing] Fix custom corpus command.
This CL changes the fuzzing documentation fixing the command that allows
users to add their own corpus to fuzzers. I had some issues trying to
get the original command to upload my corpus to the device, but after
reading the code I think it expects you to use the --staging flag to
specify the folder from where to copy the corpus to the device.

Change-Id: I7ebd2c2db023d4fbf1ae369047b1d63f2f989c38