[a11y][settings] Propagate setUI settings through settings manager
This registers a settings provider from the a11y_manager app so that setUI settings changes get propogated to settings watchers. This means that the app is both a watcher and a provider for fuchsia::accessibility::Settings, which is very confusing and will shortly be fixed as part of removing fuchsia.accessibility/settings.fidl.

The dual provider/watcher thing means that test I added in fxr/317429 still covers this functionality - instead of calling OnSettingsChanged manually, this calls settings provider methods which in turn calls NotifyWatchers which triggers the OnSettingsChanged callback.

Also makes the setui mock for this test more realistic.

Test: fx run-test a11y_tests

BUG: 16721
Change-Id: Ib9407c169a7a48919bd2cb2dbdaa7892874f2d9b