[test][clocks] Run core tests against an x64 system using HPET.
Most of the time when we are running an x64 based system, we will end up
choosing TSC as our clock source for the kernel.  This happens pretty much on
all modern hardware as well as when using KVM QEMU.

If the TSC in the system is not invariant, however, we will choose to fall back
to using the HPET instead.  This is not common, but it can happen in older
hardware, and especially in VMs.  We need to make sure that things don't
accidentally end up breaking in our support for HPET in addition to TSC.

To assist in this, add a zbi-test which passes the kernel command line argument
to force selection of the HPET, and to run the core tests.  Only do this if we
are building for x64; this argument is ignored in ARM64 kernels for obvious
reasons.

Bug: 36783
Change-Id: Id36f177414fad529574504c25a3291be4deb58bb