[DriverFramework][suspend] Fallback to reboot by default
Reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/307892
which was reverted due to test failures that are resolved here
https://fuchsia-review.googlesource.com/c/fuchsia/+/322505.
Also change to use lib-inspector.

When suspend task is stuck, currently we wait forever.
This is hindering infra workflows and so we should be rebooting
if the suspend timesout after 30sec. For debugging the driver suspend
issues, before kicking off the reboot, we dump the thread info of the
offending process.

BUG:fxb/34577 #Done

Test: Added automated test for suspend fail.
runtests -t devcoordinator-test. On Astro, I manually failed block
driver suspend, and ensured that the debug info is printed, with the
callstacks and dso list of the offending devhost.

Change-Id: I6f73c80bd83abbb2bbca545a03cf22b9269b54ec