[ledger] Use a counter in BackgroundSyncManager and DiskCleanupManager
The code was using a boolean to decide whether request were currently in
progress. Given that operation can be asynchronous (especially once
AutoCleanableSet will only delete element asynchronously), this class
needs to handle multiple operation at the same time and the sentinel
must be a counter instead of a boolean.

Change-Id: I96e19471f6a79fcd2cc8882bec83b2aa62856b4d