[exception_broker] Expanded ProcessLimbo interface.
It now also lets the user query what exceptions are available, and have
another way of actually retrieving them.

This is a better semantic, as the caller of limbo might not be
interested in all the exceptions and by the time it received the
exception handle, it is already too late.

TEST=Expanded unit test.
     fx run-test exception_broker_tests

Change-Id: I6a34a829346295d0d213fcf50436508bb83f0669