[inspect][validator] Report unimplemented acts
Report when a puppet returns not_implemented for an action.

Add a couple of pretty-printing options.

Remove overly complicated data structures: TrialSet and Quirks.
(Quirks will return, but not tied to the Trial data structure.)

Switch from StructOpt to Argh.

Test: README.md
Also, when run against the Dart puppet (not checked in at
this time), it reports correctly that CreateProperty(Uint) is unimplemented.
CF-911

Change-Id: I8cdb13708ad0e0c71c072ad21b8a43244c6e248f