[Clang][Fuchsia] Temporarily disable -Wint-in-bool-context warnings
Clang introduces a new warning, ported from GCC, which warns when
integers are used in instances where booleans are used. Example:

```
bool b = (1 << 3);  // converting the result of '<<' to a boolean; did you mean '(1 << 3) != 0'?
```

This seems odd since conversion from ints to bools I think are already defined
in C++. Regardless, since multiple instances of this warning appear when
building zircon, we should silence it for now then revisit it once the
toolchain is rolled and developers won't add new instances this can
appear. This warning was silenced in zircon but appears in fuchsia as
well.

Bug: 37215

Change-Id: I19a363eef698377f8a8ba51f8274636330a0ddf3