[wlan][ralink] Support SME channel
BUG: WLAN-1244 #done #comment

Fuchsia is no longer committed to maintaing its Ralink driver.
In fact, the driver is considered deprecated and is in the
process of being deleted entirely: go/wlan-ralink-deprecation
Deprecating the driver is a process. We stopped maintaining
the driver a while ago and do not intend to spend resources
on bringing up a test framework for it. At the same time,
not transitioning the driver to a new WLAN driver feature
will prevent the team from making progress on other construction
sites. Hence, the driver is moved to the new feature.

Bug: 29857
Test:
Automated smoke tests on Ralink in our MTV lab.
Manual verification with Ralink for Client and AP mode.

Change-Id: I4b0185b7815dab5fc25c6b12f73f88741b7bb294