[runtests-utils-test] Disable as it is flaking
This test has been added here https://fuchsia-review.googlesource.com/c/fuchsia/+/315407
but it started flaking frequently recently. The flake is an assert in async lib.
From recent experience of fixing the same assert, my guess right now, is that the test is
somehow using async loop incorrectly. This is majorly blocking rollers, and so I have resorted to
disabling the test.

Bug: 37248

Change-Id: I2d8362d193329e9d68470ea30c7ec5897315df8b