[audio_core] Add ObjectRegistry interface.
Define a new interface for a large chunk of AudioDeviceManagers API
footprint. This interface is sufficient to cover the functionality
needed by AudioDevice and derivitives.

This will make it straight forward to provide a simple/fake
implementation for testing purposes. We also begin adding a few basic
fakes for ObjectRegistry and AudioDevice sufficient to support the
existing ReporterTest. These are very basic for now but the intention
is these will be extened over time as tests require more
sophisticated behavior from these fakes.

Test: No behavior change; only a refactor.
      fx run-test audio_core_unittests
Change-Id: I639b660909591b42500693d3ab8bf027816d5ea8