[zircon_benchmarks] componentize and shellify
Bare tests are being deprecated, and so is /pkgfs/packages access.
We move the test validation path to be a component, and the manual invocation
path to be a shell binary, which keeps the user-facing behavior equivalent to
how it is run today.
The test case should drop deprecated sandbox features in the future.

Bug: 23010
Change-Id: I2410b11b2a11bb9879fe57ef1f0b453431e088e4