[media][tests] Be explicit when testing resamplers
When testing the audio resamplers, we should not depend on default
heuristics that determine which resampler is selected.  Instead, we
should explicitly select the specific resampler being tested.  This
was already happening in almost all cases, but with this CL we are
100% consistent across-the-board, including not using the enum
Resampler::Default when one isn't specified.  (We also found/fixed
a few cases where the incorrect resampler was specified.)

Test: test-only change, no product behavior change

Change-Id: I6e9789f1f0143cf05c9fd7576ccd910c7b19112b