[kernel][arm64] Use x15 for percpu pointer rather than x18
x18 is reserved for the shadow-call-stack ABI.  LLVM doesn't
allow making x16, x17, or x19 fixed, for its own reasons.  x20
and above are not saved and restored for the "short" exception
path.  So x15 is the natural choice, though one of x19..x28 would
be more ideal in theory because they're callee-saved in the ABI
and so the added register pressure from reserving one doesn't
increase the number of register saves/restores required.

Bug: 34451
Change-Id: I0ec7cd858981171540f1181dd9d77444ea081d50