[zbl] Warn on use of fbl::AutoLock without a name.
The code construct:

  {
    fbl::AutoLock(&my_lock);
    // perform action in critical section
  }

presumably intends to hold "my_lock" over a scope, but --- because the
fbl::AutoLock was not given a name --- actually just acquires and
immeadiately releases "my_lock" in the first line.

While the fix is simple (use "fbl::AutoLock lock(&my_lock)"), the
difference is subtle enough to be missed.

This CL adds "[[nodiscard]]" to the constructors of fbl::AutoLock,
causing the compiler to emit a warning if a fbl::AutoLock temporary is
created without being used.

We also fix up two instances in "system/dev/misc/pty/pty-core.cc" and
"system/ulib/fdio/unistd.cc" where fbl::AutoLock is being misused.

Change-Id: I0df47943c5dc4a63006bbae5ca5e0507ce6e1e99