[lockdep] Warn on use of Guard without a name.
The code construct:

  {
    Guard(&my_lock);
    // perform action in critical section
  }

presumably intends to hold "my_lock" over a scope, but --- because the
Guard instance was not given a name --- actually just acquires and
immeadiately releases "my_lock" in the first line.

While the fix is simple (use "Guard lock(&my_lock)"), the
difference is subtle enough to be missed.

This CL adds "[[nodiscard]]" to the constructors of Guard,
causing the compiler to emit a warning if a Guard temporary is
created without being used.

We also fix up two instances of Guard being used in "kernel/mp.cc"
where the Guard was being misused.

Change-Id: Ia672a2a7369a33b1a66031d923edf349aef22301