[kernel] Warn on use of AutoSpinLock without a name.
The code construct:

  {
    AutoSpinLock(&my_lock);
    // perform action in critical section
  }

presumably intends to hold "my_lock" over a scope, but --- because the
AutoSpinLock instance was not given a name --- actually just acquires and
immeadiately releases "my_lock" in the first line.

While the fix is simple (use "AutoSpinLock lock(&my_lock)"), the
difference is subtle enough to be missed.

This CL adds "[[nodiscard]]" to the constructors of the auto lock,
causing the compiler to emit a warning if a temporary is
created without being used.

Change-Id: I353b7df3f75363ef4bc965be95e9af36473a9b34