[kernel][debuglog] Don't terminate early during shutdown
This change eliminates a very unlikely race condition.

The kernel's dlog dumper thread is responsible for reading debuglog
messages and writing them to the kernel console and serial
port. dlog_shutdown() signals the dumper thread that the debuglog
subsystem is being shut down and waits for the the dumper thread to
flush any outstanding messages before terminating.

Prior to this change, depending on timing, it was possible for the
dumper thread to terminate before processing all messages written
prior to dlog_shutdown(). For that to happen, a new message must be
logged and shutdown must be signaled just *after* the dumper has
finished its dlog_read loop. On the next iteration, the dumper would
check the dumper_shutdown_requested variable and break early before
performing one last dlog_read loop.

Although the race condition is unlikely to occur in practice, this
change eliminates it by executing the dlog_read loop after recognizing
that shutdown has been requested.

Test: manually shutdown repeatedly, observe "Shutting down debuglog"
Change-Id: Ice133b1beddb077cb561f60d43b0a467059e065d