[debugger] Skip prologues when stepping into.
Makes the "step-into" ("s") command automatically skip function
prologues when stepping into a new function. This is desirable so that
the stack and function parameters can be interrogated properly by the
user when the step complets.

Some additional test infrastructure is added to support different
MockModuleSymbols implementations for the thread controller test, and to
allow derived classes of this test. This is used to provide a
prologue-aware mock of the symbol system for the new test.

BUG=5372

Change-Id: I77e0d0d558f25bdaaeafb33efcba1dd7706fc7c0