[debugger] Skip prologues when stepping into.
Makes the "step-into" ("s") command automatically skip function
prologues when stepping into a new function. This is desirable so that
the stack and function parameters can be interrogated properly by the
user when the step complets.

Some additional test infrastructure is added to support different
MockModuleSymbols implementations for the thread controller test, and to
allow derived classes of this test. This is used to provide a
prologue-aware mock of the symbol system for the new test.

BUG=5372

Reland of 8c9241ff3eb04aba8d2f13585770794647a39ed6 with merge fix.

Change-Id: If88ff88d4c4d3be63133c212cd3751bff6d121e7