[scenic][a11y] Transforms a11y pointer event listener into fidl event model.
previously, the interface and a11y input logic assumed that not every
fidl callback needed to be invoked. The fidl interface and implementation then
invoked a callback only in specific points for synchronizing, dropping the rest.

This is not allowed.

This change then implements this synchronization mechanism through a FIDL event.

Please see:
http://go/fxb/36059

TEST: fx run-test a11y_lib_tests
    fx run-test a11y_tests
    fx run-test scenic_tests -t input_unittests
Change-Id: I1df54afe4b20930ff6aab38e2679465ef824d260