[ledger] Implement eager garbage-collection policy
This commit implements an eager garbage-collection policy that attempts
to delete an object whenever it does not have any live reference.

It has the following limitations, that we intend to fix in follow-up
commits:
- it performs a lot of disk reads, since any object that stops being
  used will trigger database reads to check its on-disk references.
- objects that fail to be garbage-collected before a page is closed (eg.
  unused PutReference results, or upon unexpected shutdown) are not
  scanned and collected later.

It is however very effective at finding garbage-collection-related bugs
in other parts of the code, which is why we want to land it as soon as
possible (and potentially keep using it in the future for tests).

Bug: 12124
Change-Id: Ifd22ec0b2c4448e5b9b4ad46396f5b3dade20027