[tools][lib][cache] Fix flake
Because of how part of the filecache test waits to access a file
it can rare occur that a thread waits longer than the lifetime of
the rest of the fuzzer but no as long as it takes to exit causing it
to try and read a file after the cache has been deleted. Adding to
the existing wait group solves this issue however.

Change-Id: I34bc9cbba8c4e7757302bf048bb9385c4695eb4d