[syscalls][kazoo] Implement json generator
This is the last backend (and turned out to be by far the trickiest)
however! All outputs are identical* now.

The asterisk on "identical" applies to the json output. The only
consumer that I can find of the json output is the doc updater which
doesn't care about these two things:
1. The array sizes (they aren't specified in fidl, and weren't used
   except by the json generator)
2. The argument IN/OUT/INOUT/optional specs. These are implicit in
   fidl-kazoo-format and so there were a lot more that got OUT put on
   them, than were manually specified in syscalls.abigen. It doesn't
   seem too useful to add those to syscalls.abigen at this point just to
   make the output match.

This also moves all the prototype syntax from
  [In0=":rights:READ blah blah"]
to simply part of the doc comment for the method. This allows for
matching the current banjo/abigen output as we haven't decided on the
syntax to use in the fidlc front end, so there's not much point in
trying to make up a new syntax and then parse it, rather than just
eventually implementing it in fidlc properly. It's also less ugly to
read this way.

Adds tests for the json backend too.

Bug: 33668
Change-Id: Id45112be3cf55a45e6b6e429148606b7808a948f