[fx] add debug log option to fx build
Manually tested the following scenarios:
`fx build` with and without a compilation error
`fx build -k 0` with and without a compilation error
`fx build --log LOGFILE` with and without a compilation error
`fx build --log LOGFILE -k 0` with and without a compilation error
all of the above with and without a landmine file change (gn gen re-run)

If Ninja fails and no --log option is present, a "hint" message is
printed with instructions to use "--log" and attaching the result if
reporting a bug.

If Ninja fails and the --log option is present, a warning message is
printed with the location of the log file and asking the user to attach
it to the bug if reporting a bug.

If Ninja succeeds, nothing is changed in the output.

Bug: 7115

Change-Id: I76a3e5b85f8c2ab6b7c6b647894863de07687cbf