[artifactory] Stop trying to update objects if seen
This requires the user to have deletion permissions, which is something
we cannot give the bots that run untrusted code for security reasons.
A follow-up design  will lay out a secure scheme for updating timestamps for more reliable garbage collection.

Note: the affected lines pertain to internal details of the
'real' implementation of an interface that is passed into thoroughly
tested functions. Moreover, the tool was tested locally in uploading to
gs://fuchsia-infra-test.

Change-Id: I85657772ce6ba8cbe5a272ee41ab94ebe77f3f45