[debugger] Search local names for locations.
Uses the new infrastructure for finding names to match breakpoints on
local function definitions. If you type "break Foo" or another command
that takes an input location, "Foo" will match "MyClass::Foo" when the
current location is inside a member of MyClass.

It will also match global versions of "Foo" as before, so this new
behavior is additive. To implement this, many places that took an
InputLocation now take a vector. But very little needed to change at the
low level because one InputLocation could already match multiple
addresses.

Separates out MockThread from the frame_impl_unittest.cc so it can be
used by the new InputLocationParser tests.

Change-Id: I946b881fdb373c4ce21c2f55bc9770da9952dc79