[media.audio] Load provided volume curve
This loads a provided default volume curve if it is provided to the
component. Otherwise we resort to our other default.

This also changes the generated schema string in the header to be
char[], so that it does not require runtime initialization and
cannot be incorrectly read before its initialization.

Bug: 37298
Bug: 35394
Bug: 35392

TEST: Updated; fx run-test audio_core_unittests
Change-Id: Ib7b016719c8e72a79cdc8bfad51b089dcce8f5a0