[media] Extend decryptor adapter test timeouts
When running in QEMU with the asan variant enabled, the
decryptor_adapter_tests were flaky. The symptom was that test failed to
acquire the input buffer collection from sysmem and would timeout. The
timeout was likely marginal, as it was set to the default of one second
and successful runs of these tests on the same configuration would
typically take 500-800 milliseconds. This CL increases the timeout to a
large value (15 seconds) to provide plenty of additional margin.

Bug: 34881

Change-Id: Idfdb407a77302578e6b41963f09e8849dd01e202