[netstack] filter: add rule validation to FIDL struct conversion
This CL makes the FIDL struct conversion functions to validate
the rule. Currently it detects invalid port ranges and mixed IP
versions existing in a single rule.

This validation also prevents a FIDL call from setting a bad rule
to the filter system as the bad rule is detected while the rule is
converting into the filter rule's internal form.

Bug: 35844
Change-Id: I0c808f17a064bef027026f57d5f2b3b8772924e8