[zircon][platform] Zero-initialize sys_device_proto
The test /boot/test/sys/ddk-power-test -f PowerTestCase.SystemSuspend
was failing when trying to do a clang roll. This was because at some
point we ended up calling zx_protocol_device_t::suspend_new which was
0xaaaa... indicating uninitialized data. Tracing it back to the source
led to sys_device_proto which did not initialize the suspend_new field.

This helps unblock the next clang roll and fixes this failing test on
the bringup CQ bots when attempting the roll.

Bug: 35964
Change-Id: I0e3207274d9872328c90a38126f9a052a889067c