[kazoo] Remove dependency on fxl/fbl other than for gtest
In order to move kazoo into Zircon, strip the dependency on fxl. Also
drop fbl while we're at it because it was only very lightly used.

It's still using //src/lib/fxl/test:gtest_main which will be removed in
a follow switch from gtest to zxtest.

No intended change in behavior.

Bug: 33668
Bug: 37505
Change-Id: I408893bbbdc30165a38a0e215e1e4e37203759f9