[debugger] Clean up test setup usage.
ProcessSymbolsTestSetup now has a new InjectMockModule method that does
several steps that were previously duplicated in every test. Tests using
the old way are now updated with a slight increase in clarity.

Many of the tests made a "root" variable. This name was sufficiently
vague that almost every use had a comment that this was the index root.
These variables are now renamed "index_root" and the comments removed.

TEST=this is a change to test setup only.

Change-Id: I4d3c956a423b84d4504a4c5ea63baedabf8fd88a