[cobalt] Add config data to control the environment.
https://fuchsia.dev/fuchsia-src/development/config_data

Config data, for now, only includes the environment Cobalt should
connect to. Options are DEVEL or PROD. DEVEL is the default, and is
what is used in the core/default build for now. But it will be
changed to PROD once we have the prod pipeline running and stable.

The environment is used to determine:
- which shuffler public key is used for encrypting envelopes
- which analyzer public key is used for encrypting observations
- which Clearcut log source the observations are sent to

Cobalt developers can override the default config with an argument to fx set:

fx set [product].[board] --args='cobalt_environment="DEVEL"'

Testing: integration test cobalt_testapp_no_network covers the configuration
data's use in the default environment with the cobalt_app.

Change-Id: Id28230074f34ceab9bf509ad1a83c791e3751976