[netstack3] Future-proof ARP for multiple link device protocols
Previously, our ARP implementation was brittle in that it would not
have been able to support multiple different link device protocols
all using the same hardware address type. This change fixes that,
and adds documentation to explain the problem and the solution (see
the comment added in `device/arp.rs`). NDP has the same problem; I
will follow up with a fix in a future commit.

Test: No behavior change

Change-Id: I5ac6ecd2048ed906abaf6dd84d7e104dac60e713