[docs] Add diagrams to realms.md
Add diagrams that illustrate realms and collections and explain them.

This change also introduces some improvements to the existing capability
routing diagram. The `root` realm is renamed to `shell` to be less
abstract and so as not to suggest that it corresponds the root of the
component instance tree. `shell` and `system` are also renamed to
`tools` and `services` respectively.

Change-Id: Idb0df2823673db5ab781582fea2bcb16a0750da2