[intel-i2c] Assert that device_remove succeeds
If device_remove fails then the driver is in an inconsistent state and
it's not worth continuing on. In this case the driver will avoid freeing
some memory and continue to execute, which will leak. In the future
device_remove (or device_async_remove) might not have a return value and
do its own panic/error handling.

Change-Id: I22a6e08550157b5ef5e900675dc81840ad3579c0