[componentmgr] Clarify Hooks terminology and simplify adding events
Hooks was an overloaded term prior to this CL. It was used in two
capacities:

1. Hooks was used to specify the Event such as "BindInstance", "StopInstance", etc.
2. Hooks was used to specify the object that "listened" to the event.

This CL separates the two concepts: Event and Hook.

This CL also refactors hooks.rs to look more like a traditional event dispatch system.
You construct an event object, and pass it into hooks.dispatch(&mut event).await;

Events propagate up as expected. In the future it might be possible to specify
custom events via Box<Any>. This is useful perhaps for runner events or other events
coming from other components.

Change-Id: I3cba0b3bfc7664ff72383bbec035993276221d36