[appmgr] Add new global-data feature.
This is an internal features usable only by the archivist to host
storage stats. This should be the component manager's responsibility
when it is ready.

There are two checks to prevent unauthorized access of this feature:

1. //build allowlist which adds new '__internal_features' key to control
access to the feature. This key is invalid during manifest validation so
developers cannot add it manually so long as they run manifest
validation.

2. Explicit allowlist in appmgr to restrict access to only authorized
component URLs. Even if someone overrides the build rules, they will be
unable to use the feature (verified by restricted_global_data) test.

Change-Id: I8da69e3a9b5dc4cadc7e36d9247fbeae0cb2ed23