[intel-hda] Fix race in WaitCondition preventing the full time being used.
The timeout for "WaitCondition" should be a lower bound. That is, when
"WaitCondition(/*timeout=*/ZX_MSEC(10), ..., condition)" is called, we
should guarantee that we evaluate "condition" at least once after 10
msec.

The rationale here is that if a hardware spec states, "bit X should
transition from 0 to 1 in no more than 10 msec", writing:

  WaitCondition(ZX_MSEC(10), ..., check_bit);

should not time out if the hardware is behaving.

In the current implementation however, if we are preempted at an
unfortunate time, we may call the condition at a time before the
timeout, read the clock again, and then decide we are now past the timeout
without evaluating the condition again. This may cause spurious timeouts
when WaitCondition fails to recheck the condition at the end of the
timeout period.

This CL updates WaitCondition to always call cond() at least once after
the timeout.

We additionally modify WaitCondition to use an abstract Clock to allow
testing.

Change-Id: I400b157d5e9b637b64438e38b6ef9bb1a5bbef3b