[debugger] Mark symbols fully qualified.
When retrieving a symbol's identifier, they now have a global
qualification so that the resulting names are unambiguous. Otherwise, if
we take that string and look it up, it may match other symbols in the
current scope.

It is undesirable to see "::" prefixing everything in the debugger UI so
some additional options and name retrieval functions are added to
control whether any global scoping is included in the output. In most
cases this is obvious from context so the "::" preceeding a name can be
omitted.

Change-Id: I83095fe4cf1a62902acc8209599d6b199f6b7247