[ledger][p2p] Skip the handshake.
No need for devices to exchange an handshake to exchange their IDs:

* When a device A attempts to connect to a device B, B assigns a
random ID for the incoming connection.

* When a device C sees a device D that exposes the right
service and with an ID that is greater, C attempts to connect to D.

--

The tests are now templated with a boolean to exercise both code paths:
the one with an inbound connection, and the one with an outbound
connection.

Bug: 36107

Change-Id: I81247094bdf76afb6ce1bcf8ca6eff08ab4ee7a5