[fidlgen][rust] Use Option<Box<T>> for nullable xunions
Currently the Rust bindings represent nullable xunions as Option<T>
where T is a Rust enum. This CL changes fidlgen to emit Option<Box<T>>
instead, like it already does for nullable structs and unions. This
prevents an infinite-sized type compilation error when a xunion contains
contains itself recursively, as in this example:

    xunion X { S s; };
    struct S { X? x; };

Test: fx run-test rust_fidl_tests # new unit tests in encoding.rs
Test: fx set terminal.x64 --with //bundles:kitchen_sink && fx build
Test: repro steps in fxb/36698 no longer cause compilation failure
Test: further tests to be added in GIDL with fxb/36424, tested in Rust with fxb/36400

Bug: 36698
Change-Id: I10b939c2359ae6b7d5a84e984e9db92fbdcfc372