[fidl] Improve docs about thread-hostility
These docs were written when an InterfacePtr could be bound only to the
current thread. Now that you can bind to an arbitrary
async_dispatcher_t, the docs are slightly incorrect. Rather than talking
about the thread *on* which the InterfacePtr was bound (i.e., the thread
that calld the Bind() method), the InterfacePtr has affinity for the
thread *to* which the InterfacePtr was bound (i.e., the thread
assoicated with the async_dispatcher_t that was passed to the Bind()
method).

Tests: No behavior change. Just documentation.
Change-Id: I26c286db1f7556882495c1267629f1d633d9621b