[fuchsia.boot] A protocol for getting the root job with limited rights.
There are some components which essentially just want to inspect all
processes running on the system. Today components do this by grabbing
the root job with all rights (including things like write and read)
and descending the task hierarchy. This service will still vend out the
root job, but only give the inspect, enumerate, duplicate and transfer
rights, so as to make that operation safer.

Bug: 25045
Change-Id: Ife3b46d504d91d300d4f20ee0869eb05aa0b47c9