[settings] Fix audio in setui client
Previously it was not possible to watch the audio settings from the setui client.
The client logic checked for an empty AudioSettings, but it will never be empty
since we assign values to audio_settings.streams regardless of whether the fields
are empty. This leads to a panic when the empty settings is converted.

Bug: 37506
Test: fx run-test setui_client_interface_test -- --test
Updated the watch test case to accurately reflect the case in which a
watch would be detected by the command method.

Change-Id: Ic7e5eb5663ac49c675b6872d598762effa7c73a1