[camera] CameraManager2 basic implementation.
This CL adds a very basic form of the Camera Manager
that is compatible with the fuchsia::camera2 apis.

This camera manager only connects to one device,
only allows one client, and only serves one stream.
Only partial api coverage is present.

Test: Included is a test that uses the virtual camera
to test connecting to a stream and streaming a few frames.

MULTIPLY: `[
  {
    "target": "//src/camera:camera_full_test",
    "total_runs": 100
  }
]`

Change-Id: I1235f70c4a5f494ed89d181dab0d0264f0df0016