[ledger] Avoid recursive call in ScheduleObjectGarbageCollection
ScheduleObjectGarbageCollection could (in principle) overflow the stack
because it performs unbounded recursive calls. It does not happen in
practice because the recursive call is hidden in a callback and the
intermediate DeleteObject call systematically yields by virtue of our
current implementation of LevelDb, but this is fragile.

The new code has the added advantage of not allocating a separate
coroutine for every reference recursively garbage collected.

Test: fx run-test ledger_tests -t ledger_unittests
Change-Id: I955145c8785509cddefcdc5a1880aaf4ed41e8ce