[componentmgr] Shutdown and Destroy don't resolve children
`Shutdown` and `Destroy` no longer propagate from a child realm that's
not resolved. Previously, this caused gratuituous resolution and led to
infinite recursion if two component instances in the path to root shared
the same component declaration. If a realm is not resolved, there's no
specific work required for its children, so the resolution was
unnecessary anyway.

The change also includes some refactoring to move ActionSet under its
own lock. This makes it possible to register actions on a realm that
isn't resolved. This also means that `register_action` no longer
transactionally mutates any state, aside from the action set.

Bug: 37283
Change-Id: I0e9f89a235bfe432a53f5633165c4d9bf9c135a0